Have the pins read on rising edge of the clock
Upon examination of the communication with a digital analyser, it looks like the state of the data pin should be read on the rising edge of the clock. It is reasonable to think the same is applicable to the command pin. Making this change to the code greatly improves performance and the delay setting can be set as low as 2 microseconds using an official Sony PSX controller at 4.7 Vcc, which translates to a clock speed of around 40 KHz.

This could be much faster though, according to https://store.curiousinventor.com/guides/PS2